+++ src/Configuration.php
@@ -41,7 +41,7 @@ class Configuration
 
     /**
      * @param string|null $key
-     * @param mixed|null $default
+     * @param mixed|null  $default
      *
      * @return mixed|null
      */